memberService 기능 통합
