+++ src/main/java/msg/team1/Hi/domain/member/controller/MemberController.java
@@ -5,8 +5,7 @@ import msg.team1.Hi.domain.member.dto.request.ChangePasswordRequest;
 import msg.team1.Hi.domain.member.dto.request.LoginRequest;
 import msg.team1.Hi.domain.member.dto.request.SignUpRequest;
 import msg.team1.Hi.domain.member.dto.response.MemberLoginResponse;
-import msg.team1.Hi.domain.member.service.ChangePasswordService;
-import msg.team1.Hi.domain.member.service.MemberService;
+import msg.team1.Hi.domain.member.service.impl.MemberService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +18,6 @@ import org.springframework.web.bind.annotation.*;
 public class MemberController {
 
     private final MemberService memberService;
-    private final ChangePasswordService changePasswordService;
 
     @PostMapping(value = "/signup", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> signUp(@RequestBody @Validated SignUpRequest signUpRequest) {
@@ -35,7 +33,7 @@ public class MemberController {
 
     @PatchMapping("/password")
     public ResponseEntity<Void> changePassword(@RequestBody @Validated ChangePasswordRequest changePasswordRequest) {
-        changePasswordService.execute(changePasswordRequest);
+        memberService.changePassword(changePasswordRequest);
         return ResponseEntity.noContent().build();
     }
 }
