--- src/main/java/msg/team1/Hi/domain/member/service/ChangePasswordService.java
@@ -1,7 +1,0 @@
-package msg.team1.Hi.domain.member.service;
-
-import msg.team1.Hi.domain.member.dto.request.ChangePasswordRequest;
-
-public interface ChangePasswordService {
-    void execute(ChangePasswordRequest changePasswordRequest);
-}
